+++ src/modify/omgrod.geometry-dash-surge/GDSIslandSelectLayer.cpp
@@ -1,0 +1,7 @@
+#include <Geode/Geode.hpp>
+#include "../../SwelvyBG.hpp"
+#include <alphalaneous.alphas_geode_utils/include/NodeModding.h>
+
+using namespace geode::prelude;
+
+SET_SWELVY_SPRITE(GDSIslandSelectLayer, "omgrod.geometry-dash-surge/GDSIslandSelectLayer");